+++ TM4C123_ACOMP/xUtils/DataStructure/LinkedList/CircularSingleLinkedList/xSource/CSLinkedList_Node.c
@@ -88,7 +88,7 @@
      uint32_t u32SizeReg = 0UL;
      uint32_t u32SizeList = 0UL;
 
-     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != pvData ) && (0UL != u32MaxLength))
+     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != (uint32_t)pvData ) && (0UL != u32MaxLength))
      {
          psMember = CSLinkedList__psGetHead(psList);
          u32SizeList = CSLinkedList__u32GetSize(psList);
@@ -114,7 +114,7 @@
      uint32_t u32SizeReg = 0UL;
      uint32_t u32SizeList = 0UL;
 
-     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != pvData ) && (0UL != u32MaxLength)&& (0UL != u32Nodes))
+     if(((uint32_t)0UL != (uint32_t)psList) && ((uint32_t)0UL != (uint32_t)pvData ) && (0UL != u32MaxLength)&& (0UL != u32Nodes))
      {
          psMember = CSLinkedList__psGetHead(psList);
          u32SizeList = CSLinkedList__u32GetSize(psList);
