+++ TM4C123_ACOMP/xUtils/DataStructure/Set/xSource/Set_Union.c
@@ -66,7 +66,7 @@ Set_nSTATUS Set__enUnion(Set_TypeDef* psNewSet, Set_TypeDef* psSet1, const Set_T
                 {
                     pvDataMember = SLinkedList__pvGetElementData((const SLinkedListElement_TypeDef*)psMember);
                     psMemberSet = Set__psIsMember(psSet1, pvDataMember);
-                    if((uint32_t)0UL == psMemberSet)
+                    if((uint32_t)0UL == (uint32_t)psMemberSet)
                     {
                         psNewMember = SLinkedList__psAddEnd((SLinkedList_TypeDef*) psNewSet,  pvDataMember);
                         if( (uint32_t)0UL == (uint32_t)psNewMember)
@@ -124,7 +124,7 @@ Set_TypeDef* Set__psUnion(Set_TypeDef* psSet1, const Set_TypeDef* psSet2)
                 {
                     pvDataMember = SLinkedList__pvGetElementData((const SLinkedListElement_TypeDef*)psMember);
                     psMemberSet = Set__psIsMember(psSet1, pvDataMember);
-                    if((uint32_t)0UL == psMemberSet)
+                    if((uint32_t)0UL == (uint32_t)psMemberSet)
                     {
                         psNewMember = SLinkedList__psAddEnd((SLinkedList_TypeDef*) psNewSet,  pvDataMember);
                         if( (uint32_t)0UL == (uint32_t)psNewMember)
